LayoutTests/imported/w3c/web-platform-tests/css/css-ruby/inheritance-expected.txt


PASS Property ruby-align has initial value space-around
PASS Property ruby-align inherits
FAIL Property ruby-position has initial value alternate assert_equals: expected "alternate" but got "over"
PASS Property ruby-position inherits
FAIL Property ruby-merge has initial value separate assert_true: ruby-merge doesn't seem to be supported in the computed style expected true got false
FAIL Property ruby-merge inherits assert_true: ruby-merge doesn't seem to be supported in the computed style expected true got false
PASS Property ruby-overhang has initial value auto
PASS Property ruby-overhang inherits
